refactor : menu- product 패키지 병합
