+++ src/main/java/kitchenpos/menu/application/ProductService.java
@@ -1,8 +1,8 @@
-package kitchenpos.product.application;
+package kitchenpos.menu.application;
 
 import java.util.List;
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.repository.ProductRepository;
+import kitchenpos.menu.domain.Product;
+import kitchenpos.menu.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
