+++ src/main/java/kitchenpos/menu/domain/MenuPrice.java
@@ -4,7 +4,6 @@ import java.math.BigDecimal;
 import java.util.Objects;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
-import kitchenpos.product.domain.ProductPrice;
 
 @Embeddable
 public class MenuPrice {
