+++ src/main/java/kitchenpos/menu/dto/MenuProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.dto;
+package kitchenpos.menu.dto;
 
 import kitchenpos.menu.domain.MenuProduct;
