+++ src/main/java/kitchenpos/menu/dto/MenuProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.dto;
+package kitchenpos.menu.dto;
 
 import kitchenpos.menu.domain.MenuProduct;
