+++ src/main/java/kitchenpos/menu/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.product.repository;
+package kitchenpos.menu.repository;
 
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
