+++ src/test/java/kitchenpos/menu/domain/ProductPriceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.domain;
+package kitchenpos.menu.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class PriceTest {
+class ProductPriceTest {
 
     @Test
     @DisplayName("가격이 음수 일 경우 에러 반환")
